+++ src/lib/litegraph/src/utils/widget.ts
@@ -1,5 +1,9 @@
 import type { LGraphNode } from '@/lib/litegraph/src/LGraphNode'
-import type { IWidgetOptions } from '@/lib/litegraph/src/types/widgets'
+import type {
+  IBaseWidget,
+  IWidgetOptions
+} from '@/lib/litegraph/src/types/widgets'
+import type { WidgetRenderState } from '@/stores/widgetValueStore'
 import type { WidgetId } from '@/types/widgetId'
 import type { UUID } from '@/utils/uuid'
 
@@ -31,6 +35,27 @@ export function getWidgetIds(
   return widgets
     .map((widget) => widget.widgetId)
     .filter((id): id is WidgetId => id !== undefined)
+}
+
+function isDOMBackedWidget(widget: Readonly<IBaseWidget>): boolean {
+  if ('isDOMWidget' in widget && typeof widget.isDOMWidget === 'boolean') {
+    return widget.isDOMWidget
+  }
+  return (
+    ('element' in widget && !!widget.element) ||
+    ('component' in widget && !!widget.component)
+  )
+}
+
+export function deriveWidgetRenderState(
+  widget: Readonly<IBaseWidget>
+): WidgetRenderState {
+  return {
+    advanced: widget.options?.advanced ?? widget.advanced,
+    hasLayoutSize: typeof widget.computeLayoutSize === 'function',
+    isDOMWidget: isDOMBackedWidget(widget),
+    tooltip: widget.tooltip
+  }
 }
 
 export function resolveNodeRootGraphId(
